Expose a FunSuite trait while preserving API
